requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1
httpx==0.26.0
loguru==0.7.2
python-multipart==0.0.6
solana==0.32.0
solders==0.20.0
base58==2.1.1
